requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
h11==0.12.0
Jinja2==3.0.1
MarkupSafe==2.0.1
pydantic==1.8.2
python-dotenv==0.19.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
